Custom_File_CompressionAndDecompression-main/Custom_File_CompressionAndDecompression-main/requirements.txt
customtkinter
pillow
PyMuPDF
opencv-python
